Package for Tileset, refactor TilesetManager
