+++ megamek/i18n/megamek/client/messages.properties
@@ -808,7 +808,8 @@ CommonSettingsDialog.pathFiderTimeLimit=Pathfinder time limit (milliseconds):
 CommonSettingsDialog.protoMechUnitCodes=ProtoMech unit codes:
 CommonSettingsDialog.rightDragScroll=By right-clicking on the map and dragging.
 CommonSettingsDialog.scrollSensitivity=Scroll sensitivity:
-CommonSettingsDialog.showDamageLevel=Show a colored damage icon next to unit
+CommonSettingsDialog.showDamageLevel=Show damage to units on the unit label
+CommonSettingsDialog.showDamageDecal=Show damage to units on the unit icon
 CommonSettingsDialog.showMapHexPopup=Show map hex popup.
 CommonSettingsDialog.showMapsheets=Show mapsheet borders.
 CommonSettingsDialog.showUnitId=Show each unit's unique Id next to its name.
