+++ megamek/src/megamek/client/ui/swing/tileset/MechTileset.java
@@ -19,7 +19,7 @@
  * Created on April 15, 2002, 9:53 PM
  */
 
-package megamek.client.ui.swing;
+package megamek.client.ui.swing.tileset;
 
 import java.awt.Component;
 import java.awt.Image;
