+++ megamek/src/megamek/test/MechSetTest.java
@@ -24,7 +24,7 @@ import java.io.StreamTokenizer;
 import java.util.Arrays;
 import java.util.Comparator;
 
-import megamek.client.ui.swing.TilesetManager;
+import megamek.client.ui.swing.tileset.TilesetManager;
 import megamek.common.Configuration;
